Use size_t for length and count variables to fix C4267
- Change wordlist_new length parameter and global counters from int to
  size_t.
- Add explicit (int) cast for WORDBUF_LEN in romaji.c to prevent MSVC
  warning C4267 while maintaining negative index checks.
- Add len > 0 guard before memcpy in wordlist_new for safety.

+++ src/wordlist.h
@@ -13,15 +13,15 @@ struct wordlist
     wordlist *next;
 };
 
-extern int n_wordlist_open;
-extern int n_wordlist_close;
-extern int n_wordlist_total;
+extern size_t n_wordlist_open;
+extern size_t n_wordlist_close;
+extern size_t n_wordlist_total;
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-wordlist *wordlist_new(const unsigned char *ptr, int len);
+wordlist *wordlist_new(const unsigned char *ptr, size_t len);
 void wordlist_destroy(wordlist *p);
 
 #ifdef __cplusplus
